examples/json_to_pptx_agent/pyproject.toml
[project]
name = "json-to-pptx-agent"
version = "0.1.0"
description = "JSON to PPTX conversion agent example"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.18",
    "langchain-core>=0.3.34",
    "langchain-openai>=0.3.5",
    "middleman-ai",
    "python-dotenv>=1.0.1",
]

[tool.uv]
[tool.uv.sources]
middleman-ai = { path = "../../", editable = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = [
    "src/json_to_pptx_agent",
]
